+++ crates/tandem-server/src/http/tests/global.rs
@@ -4,4 +4,5 @@ use std::process::Command;
 include!("global_parts/part01.rs");
 include!("global_parts/part02.rs");
 include!("global_parts/part03.rs");
+include!("global_parts/part05.rs");
 include!("global_parts/part04.rs");
